Removed unused textures Added button textures

+++ SP4/Base/Source/PhysicsObject.cpp
@@ -13,7 +13,7 @@ CPhysicsObject::~CPhysicsObject()
 void CPhysicsObject::Update(double dt)
 {
 	//m_fRotation += dt * m_fRotationSpeed;
-	m_fPosX += dt * m_fSpeedX;
+	m_fPosX += static_cast<float>(dt) * m_fSpeedX;
 	if (m_fPosX > 0.8f)
 		m_bActive = false;
 }
